FEATURE: Make taxonomies sortable
Since all Taxonomy items are nodes, the sortingIndex of the nodes is already
respected. Instead of just having an alphabetical sorting by title, we allow
sorting taxonomies in the backend module. This is helpful if taxonomies are used
e.g. as a filter in the frontend and you explicitly want to sort the items.